+++ project/front/app/src/pages/css/begingame.css
@@ -1,0 +1,6 @@
+.gameimg {
+    width: 150px;
+    height: 150px;
+    filter: invert(1);
+  }
+
