+++ project/front/app/src/template/template.css
@@ -39,7 +39,7 @@ code {
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  height: 100vh;
+  height: calc(100vh - var(--header-size));
 }
 
 .navbar {
